use a test suite to test diff eventstores
